+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* Implemented Pixi task collection (\#812);
+
 # 1.18.5
 
 * Changed default first task selection logic (\#810);